+++ src/BDSCOD/Utility.hs
@@ -21,7 +21,7 @@ processEvent' (_, currTime) epiSimEvent =
   case epiSimEvent of
     (Infection absTime _ _) ->
       ((absTime - currTime, OBirth), absTime)
-    (Sampling absTime _) -> ((absTime - currTime, OSample), absTime)
+    (Sampling absTime _) -> ((absTime - currTime, ObsUnscheduledSequenced), absTime)
     (Catastrophe absTime (People persons)) -> ((absTime - currTime, OCatastrophe . fromIntegral $ V.length persons), absTime)
     (Occurrence absTime _) ->
       ((absTime - currTime, OOccurrence), absTime)
@@ -124,3 +124,9 @@ invLogit a = 1 / (1 + exp (- a))
 -- | Return the log-odds from the probability
 logit :: Probability -> Double
 logit p = log (p / (1 - p))
+
+
+-- | The log-sum-exp function
+logSumExp :: (Floating a, Ord a) => [a] -> a
+logSumExp xs = x' + log (sum [exp (x - x') | x <- xs])
+               where x' = maximum xs